move wip peggy files to own folder
